sonar critical to cast operand
Cast one of the operands of this multiplication operation to a long
Sonar Link:
https://sonar.onap.org/project/issues?assignees=Sahani_Naik&id=org.onap.dmaap.messagerouter.msgrtr%3Amsgrtr&resolved=false&severities=CRITICAL
Location:
src/main/java/com/att/nsa/cambria/service/impl/MMServiceImpl.java
Line No-358

Issue-ID: DMAAP-486

+++ src/main/java/com/att/nsa/cambria/service/impl/MMServiceImpl.java
@@ -355,7 +355,7 @@ public class MMServiceImpl implements MMService {
 		final long startMs = System.currentTimeMillis();
 		long count = 0;
 
-		long maxEventBatch = 1024 * 16;
+		long maxEventBatch = (long)1024 * 16;
 		String batchlen = AJSCPropertiesMap.getProperty(CambriaConstants.msgRtr_prop, BATCH_LENGTH);
 		if (null != batchlen)
 			maxEventBatch = Long.parseLong(batchlen);
